Remove `*Error.swift` files
