+++ Sources/Jira/TicketAPI.swift
@@ -61,3 +61,21 @@ extension Ticket {
 		}
 	}
 }
+
+extension Ticket.API {
+	public enum Error {
+		case authorizationEncodingError
+		case http(Int)
+	}
+}
+
+extension Ticket.API.Error: LocalizedError {
+	public var errorDescription: String? {
+		switch self {
+		case .authorizationEncodingError:
+			return "JIRA authorization token encoding failed"
+		case .http(let statusCode):
+			return "JIRA API call failed with HTTP status code \(statusCode)"
+		}
+	}
+}
